+++ .env.example
@@ -1,0 +1,61 @@
+# Example .env file for ytdlbot configuration
+
+# Number of workers (default is 100)
+WORKERS=100
+
+# Telegram app ID
+APP_ID=
+
+# Telegram app hash
+APP_HASH=
+
+# Telegram bot token
+BOT_TOKEN=
+
+# Owner ID or username
+OWNER=
+
+# List of authorized users (comma-separated)
+AUTHORIZED_USER=
+
+# MySQL Data Source Name
+MYSQL_DSN=
+
+# Redis host
+REDIS_HOST=
+
+# Enable FFMPEG for video processing (True/False)
+ENABLE_FFMPEG=False
+
+# Desired audio format (e.g., mp3, wav)
+AUDIO_FORMAT=
+
+# Enable Aria2 for downloads (True/False)
+ENABLE_ARIA2=False
+
+# Path to Rclone executable
+RCLONE_PATH=
+
+# Enable VIP features (True/False)
+ENABLE_VIP=False
+
+# Payment provider token
+PROVIDER_TOKEN=
+
+# Free downloads allowed per user
+FREE_DOWNLOAD=5
+
+# Token price (default: 10 credits for 1 USD)
+TOKEN_PRICE=10
+
+# Rate limit for requests
+RATE_LIMIT=120
+
+# Path for temporary files (ensure the directory exists and is writable)
+TMPFILE_PATH=
+
+# Maximum size for Telegram uploads in MB
+TG_NORMAL_MAX_SIZE=2000
+
+# Maximum URL length in captions
+CAPTION_URL_LENGTH_LIMIT=150
