--- ytdlbot/payment.py
@@ -1,8 +1,0 @@
-#!/usr/local/bin/python3
-# coding: utf-8
-
-# ytdlbot - payment.py
-# 8/15/21 18:23
-#
-
-__author__ = "Benny <benny.think@gmail.com>"
